Rename vector_set > flat_set
